refactor(agent-skills): open import page directly with multi-select batch import
- Remove the dropdown-menu entry (unreliable under Electron titlebar drag layer);
  the 导入 button now opens the import dialog directly
- ImportSkillDialog becomes a unified batch import page:
  - 其他项目 tab: check multiple skills across workspaces, one-click
    batch import with success/skip/fail summary
  - 本地文件夹 tab: drag & drop or multi-select skill dirs / SKILL.md,
    with skip-or-overwrite duplicate strategy and result summary
- Drop the separate BulkImportSkillDialog component (logic merged in)

Made-with: Proma
